+++ library/src/test/java/com/datatorrent/lib/algo/LeastFrequentKeyMapTest.java
@@ -96,10 +96,10 @@ public class LeastFrequentKeyMapTest
     for (HashMap<String, Integer> h: list) {
       val = h.get("a");
       if (val == null) {
-        ccount = h.get("c").intValue();
+        ccount = h.get("c");
       }
       else {
-        acount = val.intValue();
+        acount = val;
       }
     }
     Assert.assertEquals("Count of a was ", atot, acount);
